# internal/api/openapi.yaml
openapi: "3.1.0"
info:
  title: TabMail API
  version: "2.0"
  description: |
    TabMail is a multi-tenant, API-first disposable email service with both
    inbound (SMTP receive) and outbound (SMTP send) capabilities.

    ## Quick Start for Bot/Automation Users

    ### Step 1: Get an API Key
    ```bash
    # Login to get a JWT token
    curl -X POST https://your-tabmail/api/v1/auth/login \
      -H "Content-Type: application/json" \
      -d '{"email":"you@example.com","password":"your-password"}'

    # Create an API Key with the JWT
    curl -X POST https://your-tabmail/api/v1/keys \
      -H "Authorization: Bearer <jwt_token>" \
      -H "Content-Type: application/json" \
      -d '{"label":"my-bot","scopes":["mailboxes:write","messages:read","send:write"]}'
    ```

    ### Step 2: Create a Disposable Mailbox
    ```bash
    # Get a suggested address
    curl https://your-tabmail/api/v1/domains/{domain_id}/suggest-address \
      -H "X-API-Key: tabmail_..."

    # Create the mailbox
    curl -X POST https://your-tabmail/api/v1/mailboxes \
      -H "X-API-Key: tabmail_..." \
      -H "Content-Type: application/json" \
      -d '{"full_address":"suggested-addr@domain.com","zone_id":"..."}'
    ```

    ### Step 3: Receive Emails (Poll or SSE)
    ```bash
    # Poll for messages
    curl https://your-tabmail/api/v1/mailbox/addr@domain.com \
      -H "X-API-Key: tabmail_..."

    # Or use Server-Sent Events for real-time
...
      -H "X-API-Key: tabmail_..."
    ```

...

...

...

...

...

...

...

...
servers:
...
security:
  - {}
  - APIKey: []
components:
  securitySchemes:
    APIKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Tenant API key
    MailboxToken:
      type: http
      scheme: bearer
      bearerFormat: TabMail-Mailbox-Token
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /auth/login
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Tenant API key (alias for APIKey)
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - BAD_REQUEST
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - CONFLICT
                - RATE_LIMITED
                - QUOTA_EXCEEDED
                - DOMAIN_NOT_VERIFIED
                - SEND_NOT_ALLOWED
                - INTERNAL
            message: { type: string }
...
    PermissionProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        can_send:
          type: boolean
        daily_send_quota:
          type: integer
          description: "0 = unlimited"
        daily_receive_quota:
          type: integer
          description: "0 = unlimited"
        max_mailboxes:
          type: integer
          description: "0 = unlimited"
        max_domains:
          type: integer
          description: "0 = unlimited"
        allowed_zone_ids:
          type: array
          items:
            type: string
            format: uuid
          nullable: true
          description: "null = all zones allowed"
        can_create_domains:
          type: boolean
        can_create_routes:
          type: boolean
        can_create_api_keys:
          type: boolean
        is_system:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    EffectivePermission:
      type: object
      properties:
        can_send:
          type: boolean
        daily_send_quota:
          type: integer
        daily_receive_quota:
          type: integer
        max_mailboxes:
          type: integer
        max_domains:
          type: integer
        allowed_zone_ids:
          type: array
          items:
            type: string
            format: uuid
          nullable: true
        can_create_domains:
          type: boolean
        can_create_routes:
          type: boolean
        can_create_api_keys:
          type: boolean
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /api/v1/admin/permissions:
    get:
      tags: [Permissions]
      summary: List all permission profiles
      operationId: listPermissionProfiles
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/PermissionProfile" }
    post:
      tags: [Permissions]
      summary: Create permission profile
      operationId: createPermissionProfile
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                description: { type: string }
                can_send: { type: boolean, default: false }
                daily_send_quota: { type: integer, default: 0, description: "0 = unlimited" }
                daily_receive_quota: { type: integer, default: 0, description: "0 = unlimited" }
                max_mailboxes: { type: integer, default: 0, description: "0 = unlimited" }
                max_domains: { type: integer, default: 0, description: "0 = unlimited" }
                allowed_zone_ids:
                  type: array
                  items: { type: string, format: uuid }
                  nullable: true
                  description: "null = all zones allowed"
                can_create_domains: { type: boolean, default: true }
                can_create_routes: { type: boolean, default: true }
                can_create_api_keys: { type: boolean, default: true }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/PermissionProfile" }
        "400":
          $ref: '#/components/responses/BadRequest'
        "409": { description: Profile name already exists }
  /api/v1/admin/permissions/{id}:
    patch:
      tags: [Permissions]
      summary: Update permission profile
      operationId: updatePermissionProfile
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                can_send: { type: boolean }
                daily_send_quota: { type: integer }
                daily_receive_quota: { type: integer }
                max_mailboxes: { type: integer }
                max_domains: { type: integer }
                allowed_zone_ids:
                  type: array
                  items: { type: string, format: uuid }
                  nullable: true
                can_create_domains: { type: boolean }
                can_create_routes: { type: boolean }
                can_create_api_keys: { type: boolean }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/PermissionProfile" }
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Permissions]
      summary: Delete permission profile
      operationId: deletePermissionProfile
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Deleted }
        "400": { description: Cannot delete system profiles }
        "404":
          $ref: '#/components/responses/NotFound'
  /api/v1/admin/users/{userId}/permissions:
    get:
      tags: [Permissions]
      summary: Get user's effective permissions
      operationId: getUserPermissions
      security: [{ BearerAuth: [] }]
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/EffectivePermission" }
        "404":
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Permissions]
      summary: Set user permission override
      operationId: setUserPermissionOverride
      security: [{ BearerAuth: [] }]
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                profile_id:
                  type: string
                  format: uuid
                  description: Permission profile to assign
                overrides:
                  type: object
                  description: Field-level overrides on top of the profile
                  properties:
                    can_send: { type: boolean }
                    daily_send_quota: { type: integer }
                    daily_receive_quota: { type: integer }
                    max_mailboxes: { type: integer }
                    max_domains: { type: integer }
                    allowed_zone_ids:
                      type: array
                      items: { type: string, format: uuid }
                      nullable: true
                    can_create_domains: { type: boolean }
                    can_create_routes: { type: boolean }
                    can_create_api_keys: { type: boolean }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/EffectivePermission" }
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Permissions]
      summary: Delete user permission override
      operationId: deleteUserPermissionOverride
      description: Remove user-specific permission override, reverting to the default profile.
      security: [{ BearerAuth: [] }]
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Override removed }
        "404":
          $ref: '#/components/responses/NotFound'
...